Creates src folder.
This commit creates the source folder. This is where most of the
project will reside. It currently has an app folder and a test
folder. The app folder is where the main app of the project goes.
The test folders has all the tests necessary for the project.
Overtime folders such as lib, docs, and utils can be added to the
src folder.